Tidy up and docs
