+++ src/contexts/ThemeProvider/defaultTheme.ts
@@ -1,4 +1,4 @@
-import { type Theme } from './types'
+import { type Theme } from '../../types'
 
 export const defaultTheme: Theme = {
   displayName: 'Default',
